CarPark/requirements.txt
asgiref~=3.4.1
backports.zoneinfo~=0.2.1
Django~=4.0
sqlparse~=0.4.2
tzdata~=2021.5
pytz==2021.3
setuptools==58.5.3
tzdata==2021.5
wheel==0.37.0
djangorestframework~=3.12.4